6.3.1: use i16 for Rect

+++ src/engine.rs
@@ -12,10 +12,10 @@ use web_sys::{CanvasRenderingContext2d, HtmlImageElement};
 
 #[derive(Deserialize, Clone)]
 pub struct SheetRect {
-    pub x: u16,
-    pub y: u16,
-    pub w: u16,
-    pub h: u16,
+    pub x: i16,
+    pub y: i16,
+    pub w: i16,
+    pub h: i16,
 }
 
 #[derive(Deserialize, Clone)]
@@ -110,10 +110,10 @@ pub struct Renderer {
 }
 
 pub struct Rect {
-    pub x: f32,
-    pub y: f32,
-    pub width: f32,
-    pub height: f32,
+    pub x: i16,
+    pub y: i16,
+    pub width: i16,
+    pub height: i16,
 }
 
 impl Rect {
@@ -155,8 +155,8 @@ impl Renderer {
         self.draw_rect(&Rect {
             x: position.x.into(),
             y: position.y.into(),
-            width: image.width() as f32,
-            height: image.height() as f32,
+            width: image.width() as i16,
+            height: image.height() as i16,
         });
         self.context
             .draw_image_with_html_image_element(image, position.x.into(), position.y.into())
@@ -263,8 +263,8 @@ impl Image {
         let bounding_box = Rect {
             x: position.x.into(),
             y: position.y.into(),
-            width: element.width() as f32,
-            height: element.height() as f32,
+            width: element.width() as i16,
+            height: element.height() as i16,
         };
         Self {
             element,
@@ -286,7 +286,7 @@ impl Image {
     }
 
     pub fn set_x(&mut self, x: i16) {
-        self.bounding_box.x = x as f32;
+        self.bounding_box.x = x;
         self.position.x = x;
     }
